refactor(Post): Renomeia a pasta da rota e atualiza a referencia das services específicas
Com isso as services de Fatos, Resolução e Hipótese passam a utilizar os métodos da base PostService
